Adding app.module back in and writing tests for app injection
