+++ src/core/angularfire2.ts
@@ -1,4 +1,4 @@
-import { InjectionToken, NgZone, NgModule } from '@angular/core';
+import { InjectionToken, NgZone } from '@angular/core';
 import { Subscription } from 'rxjs/Subscription';
 import { Scheduler } from 'rxjs/Scheduler';
 import { queue } from 'rxjs/scheduler/queue';
@@ -6,7 +6,7 @@ import { queue } from 'rxjs/scheduler/queue';
 import firebase from '@firebase/app';
 import { FirebaseApp, FirebaseOptions } from '@firebase/app-types';
 
-export function firebaseAppFactory(config: FirebaseOptions, name?: string): FirebaseApp {
+export function _firebaseAppFactory(config: FirebaseOptions, name?: string): FirebaseApp {
   const appName = name || '[DEFAULT]';
   const existingApp = firebase.apps.filter(app => app.name == appName)[0];
   return existingApp || firebase.initializeApp(config, appName);
@@ -15,29 +15,8 @@ export function firebaseAppFactory(config: FirebaseOptions, name?: string): Fire
 export const FirebaseAppName = new InjectionToken<string>('angularfire2.appName');
 export const FirebaseAppConfig = new InjectionToken<FirebaseOptions>('angularfire2.config');
 
-// Put in database.ts when we dropped depreciated
+// Put in database.ts when we drop database-depreciated
 export const RealtimeDatabaseURL = new InjectionToken<string>('angularfire2.realtimeDatabaseURL');
-
-export const FirebaseAppProvider = {
-   provide: FirebaseApp,
-   useFactory: firebaseAppFactory,
-   deps: [ FirebaseAppConfig, FirebaseAppName ]
- };
-
- @NgModule({
-   providers: [ FirebaseAppProvider ],
- })
- export class AngularFireModule {
-   static initializeApp(config: FirebaseOptions, appName?: string) {
-     return {
-       ngModule: AngularFireModule,
-       providers: [
-         { provide: FirebaseAppConfig, useValue: config },
-         { provide: FirebaseAppName, useValue: appName }
-       ]
-     }
-   }
-}
 
 export class FirebaseZoneScheduler {
   constructor(public zone: NgZone) {}
