+++ src/core/public_api.ts
@@ -1,1 +1,2 @@
 export * from './angularfire2';
+export * from './firebase.app.module';
